third_party/WebKit/LayoutTests/fast/dom/Document/createProcessingInstruction-invalid-data.html

<!doctype html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
['?>', 'foo?>bar'].forEach(function(data)
{
    shouldThrow("document.createProcessingInstruction('name', '" + data + "')");
});
</script>
</body>
</html>
