third_party/blink/web_tests/fast/dom/Document/createProcessingInstruction-invalid-data-expected.txt

PASS document.createProcessingInstruction('name', '?>') threw exception InvalidCharacterError: Failed to execute 'createProcessingInstruction' on 'Document': The data provided ('?>') contains '?>'..
PASS document.createProcessingInstruction('name', 'foo?>bar') threw exception InvalidCharacterError: Failed to execute 'createProcessingInstruction' on 'Document': The data provided ('foo?>bar') contains '?>'..
PASS successfullyParsed is true

TEST COMPLETE
